refactor: split monitoring into domain, app, infra and http layers
