+++ internal/monitoring/app/collect_integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package monitoring
+package app
 
 import (
 	"context"
