+++ internal/monitoring/app/fixtures_test.go
@@ -1,0 +1,103 @@
+package app
+
+import (
+	"database/sql"
+	"path/filepath"
+	"testing"
+
+	_ "github.com/go-sql-driver/mysql"
+	_ "github.com/mattn/go-sqlite3"
+)
+
+// useMonitoringDB points the package at a throwaway SQLite file carrying the
+// monitoring_log and changestream_statistics schemas, so the writers can be
+// exercised without touching the database tracked in this repository.
+func useMonitoringDB(t *testing.T) *sql.DB {
+	t.Helper()
+
+	path := filepath.Join(t.TempDir(), "sync.db")
+	t.Setenv("SYNC_DB_PATH", path)
+
+	conn, err := sql.Open("sqlite3", path)
+	if err != nil {
+		t.Fatalf("open temp sqlite: %v", err)
+	}
+	t.Cleanup(func() { _ = conn.Close() })
+
+	const schema = `
+CREATE TABLE monitoring_log (
+    id             INTEGER PRIMARY KEY AUTOINCREMENT,
+    logged_at      DATETIME DEFAULT CURRENT_TIMESTAMP,
+    db_type        TEXT NOT NULL,
+    src_db         TEXT,
+    src_table      TEXT,
+    src_row_count  INTEGER,
+    tgt_db         TEXT,
+    tgt_table      TEXT,
+    tgt_row_count  INTEGER,
+    monitor_action TEXT,
+    sync_task_id   INTEGER
+);
+CREATE TABLE changestream_statistics (
+    id              INTEGER PRIMARY KEY AUTOINCREMENT,
+    task_id         INTEGER NOT NULL,
+    collection_name VARCHAR(255) NOT NULL,
+    received        INTEGER DEFAULT 0,
+    executed        INTEGER DEFAULT 0,
+    pending         INTEGER DEFAULT 0,
+    errors          INTEGER DEFAULT 0,
+    inserted        INTEGER DEFAULT 0,
+    updated         INTEGER DEFAULT 0,
+    deleted         INTEGER DEFAULT 0,
+    last_updated    TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    created_at      TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    UNIQUE(task_id, collection_name)
+);`
+	if _, err := conn.Exec(schema); err != nil {
+		t.Fatalf("create schema: %v", err)
+	}
+	return conn
+}
+
+// emptyDB points the package at a SQLite file with no tables, so the writers
+// hit a missing-table error.
+func emptyDB(t *testing.T) {
+	t.Helper()
+	t.Setenv("SYNC_DB_PATH", filepath.Join(t.TempDir(), "empty.db"))
+}
+
+type statsRow struct {
+	Received, Executed, Pending, Errors int
+	Inserted, Updated, Deleted          int
+	LastUpdated                         string
+}
+
+func readStats(t *testing.T, conn *sql.DB, taskID int, collection string) (statsRow, bool) {
+	t.Helper()
+
+	var r statsRow
+	err := conn.QueryRow(`
+		SELECT received, executed, pending, errors, inserted, updated, deleted, last_updated
+		FROM changestream_statistics WHERE task_id = ? AND collection_name = ?`,
+		taskID, collection).Scan(&r.Received, &r.Executed, &r.Pending, &r.Errors,
+		&r.Inserted, &r.Updated, &r.Deleted, &r.LastUpdated)
+	if err == sql.ErrNoRows {
+		return r, false
+	}
+	if err != nil {
+		t.Fatalf("read stats: %v", err)
+	}
+	return r, true
+}
+
+func countRows(t *testing.T, conn *sql.DB, table string) int {
+	t.Helper()
+
+	var n int
+	if err := conn.QueryRow("SELECT COUNT(*) FROM " + table).Scan(&n); err != nil {
+		t.Fatalf("count %s: %v", table, err)
+	}
+	return n
+}
+
+// ---------------------------------------------------------------- row counting
