+++ internal/monitoring/app/logger_test.go
@@ -1,0 +1,15 @@
+package app
+
+import (
+	"io"
+
+	"github.com/sirupsen/logrus"
+)
+
+// quietLogger returns a logger that discards output, so tests do not flood the
+// test log.
+func quietLogger() *logrus.Logger {
+	l := logrus.New()
+	l.SetOutput(io.Discard)
+	return l
+}
