+++ internal/monitoring/domain/rowcount_test.go
@@ -1,8 +1,9 @@
-package monitoring
+package domain
 
 import (
 	"testing"
 
+	_ "github.com/go-sql-driver/mysql"
 	_ "github.com/mattn/go-sqlite3"
 )
