+++ internal/monitoring/infra/logger_test.go
@@ -1,4 +1,4 @@
-package notify
+package infra
 
 import (
 	"io"
