+++ internal/monitoring/infra/query_counter_gaps_test.go
@@ -1,8 +1,9 @@
-package monitoring
+package infra
 
 import (
 	"testing"
 
+	_ "github.com/go-sql-driver/mysql"
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/bson"
