+++ internal/monitoring/infra/query_counter_test.go
@@ -1,10 +1,12 @@
-package monitoring
+package infra
 
 import (
 	"strings"
 	"testing"
 	"time"
 
+	_ "github.com/go-sql-driver/mysql"
+	_ "github.com/mattn/go-sqlite3"
 	"github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/bson"
 )
