+++ internal/platform/slack/slack_test.go
@@ -1,4 +1,4 @@
-package notify
+package slack
 
 import (
 	"context"
